% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGMcrf-package.R
\name{crfdemo}
\alias{crfdemo}
\title{CRF Demo data examples}
\description{
A subset data prepared from Landsat 8 scene P168R060
}
\details{
The data was acquired on 2014-08-14. It is accompanied by digitized vector as training data.
}
