% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGMcrf-package.R
\docType{package}
\name{DGMcrf-package}
\alias{DGMcrf}
\alias{DGMcrf-package}
\title{DGMcrf: Directed Graphical Models: Conditional Random Fields}
\description{
USes directed graphicals models to model spatial depencies in image classification.
}
\author{
\strong{Maintainer}: Benson Kenduiywo \email{bensonkemboi@gmail.com} (0000-0002-8448-0499)

}
\keyword{internal}
