% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{give_col_numbers}
\alias{give_col_numbers}
\title{Column numbers}
\usage{
give_col_numbers(x, prefix = "col.")
}
\arguments{
\item{x}{A dataframe}

\item{prefix}{A character preceding df column numbers}
}
\value{
A dataframe with new column names
}
\description{
Assigns numbers as names to columns of a data frame
}
\details{
The function assign numbers as columns of a dataframe that is passed as an argument.
}
\examples{
## Rename the columns in an in-built data set
give_col_numbers(iris)

}
