% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{int_2_bit}
\alias{int_2_bit}
\title{Integer to bits}
\usage{
int_2_bit(x, bit.type = 8)
}
\arguments{
\item{x}{integer}

\item{bit}{type}
}
\value{
a list of integer
}
\description{
Convert integers into bit format
}
\details{
This function converts 32 bit integers and raw vector type.
}
\examples{
int_2_bit(1:10)

}
